Test dataset for dask dataframe with cat columns.
